Fix forever loop in burn_chart_data_spec
Date is not immutable which means that the "default"
vars we were creating resulted in the *same* value
being mutated.

There's a place in the code where we loop through dates
but every time we incremented the current date, it
also incremented the ending one, causing a forever
loop.

timezone-mock calls the default Date ctor and mutates
it under the hood which was exposing this bug.