Don't use ee_else_ce alias
In this case, the two versions of the component have almost nothing in
common. When using ee_else_ce here, the CE MrWidgetOptions specs fail,
since by default the EE component is loaded _instead_ of the CE one
(despite it being a spec for a CE component), and none of the expected
props are passed to it.

Now also use a dynamic import in the CE MrWidgetOptions to avoid
bundling the CE SecurityReportsApp in EE, as it's not used in the EE
MrWidgetOptions template.

Also, remove the Grouped prefix, since this component does not implement
grouping.