Speed up pull mirroring by asking Gitaly if tags have changed
Currently, pull mirroring will fetch the full list of tags for a
repository twice - before and after the pull - and use the diff between
them to work out which new tags have been added.

This seems unavoidable with the current output of `git fetch`, but we
*can* skip doing the second call (and cache invalidation) if there have
been no tag changes. This is the common, happy path, so should be worth
implementing.