Auto-assign merge request author when creating from issue
When creating a merge request from an issue via "Create merge
request" button, whoever clicked it will be the author of
the merge request.

Additional to that, we are now auto-assigning the merge request
to the author. This is to make it more convenient for the
author to assign themselves.