Fix bug when creating snippet and default branch not master
When a new snippet repository is created, the ref inside
HEAD is `master` even when the default branch is different.

This also happens with projects and wikis. These resources fix
this by explicitly rewrite the HEAD reference and we're applying
the same behavior to snippets.

When the repository is pushed or pulled we update the reference
inside the HEAD file.