Remove costly preload on pipeline schedules page
When preloading a has_one association for more than
one record, Rails loads the entire association and
selects the appropriate single record in memory.
This can cause large numbers of records to be requested
and not used.

Instead we can deliberately not preload, and although
this introduces an N+1 to fetch the associated has_one
record, it allows Rails to fetch each one individually
(with a LIMIT clause) which is much less work overall.